admin 페이지 추가 및 일부 기능 통일
