+++ src/main/java/com/lotte/cinema/board/faq/controller/FaqController.java
@@ -5,17 +5,13 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import com.lotte.cinema.board.faq.dto.FaqDTO;
 import com.lotte.cinema.board.faq.dto.FaqPageDTO;
 import com.lotte.cinema.board.faq.entity.FaqCategory;
 import com.lotte.cinema.board.faq.service.FaqService;
