+++ src/main/java/com/lotte/cinema/theater/entity/Region.java
@@ -1,0 +1,5 @@
+package com.lotte.cinema.theater.entity;
+
+public class Region {
+	
+}
