+++ src/main/webapp/WEB-INF/views/admin/admin.jsp
@@ -15,8 +15,7 @@
 	<link rel="stylesheet" href="https://cdn.ckeditor.com/ckeditor5/42.0.0/ckeditor5.css">
 	
   </head>
-
-  	
+  <body>
   	<div id="contents">
 		<div class="title_top"><h2 class="tit">Admin Page</h2></div>
 		<ul id="pageLink" class="tab_wrap outer">
@@ -27,6 +26,7 @@
 		<div id="adminContent">
 	
 		</div>
+		<a href="/customer/faq">홈페이지 이동 ></a>
   	</div>
     <script type="importmap">
 		{
@@ -37,6 +37,7 @@
 		}
 	</script>
   	<script type="module" src="/resources/static/js/admin/admin.js"></script>
+	<script type="module" src="/resources/static/js/admin/customer/customer.js"></script>
 	<jsp:include page="../commons/vendorBody.jsp"/>
   </body>
 </html>
