+++ src/main/webapp/resources/static/css/commons/content.css
@@ -24,7 +24,7 @@
 /*==== Tab-animation ====*/
 /* .area__movingbar .tab_wrap.outer > li.active > .tab_tit{ border-bottom: 0; } */
 
-
+.btn_btm_wrap{margin: 60px 0; text-align: center;}
